Further refractored with focus on strcpy() and strcat().
I've opted to leave ascii_map.c and dg_variables.c mostly as is, due to their heavy usage of strcat() and reduced readability. The goal here was to refractor and altering the code within those files without doing a deep dive would be dangerous. Refractoring those can be left for another day.